refactor: migrate from Biome to ESLint + Prettier and align project config
- Replace Biome with ESLint (strictTypeChecked) + Prettier for linting and formatting
- Add husky pre-commit hook to enforce pnpm check before commits
- Fix all non-null assertions in src with proper type guards
- Fix all ESLint errors in test files (unsafe access, require-await, unused vars)
- Rewrite test helpers to use node:assert instead of vitest expect
- Split tsconfig into src/test project references

+++ src/token/services/kms-sm.service.ts
@@ -21,7 +21,7 @@ async function getIssuerWallet(secretPath: string): Promise<Wallet> {
     throw new Error("Failed to retrieve issuer seed from Secret Manager");
   }
 
-  const resolvedSeed = typeof payload === "string" ? payload : new TextDecoder().decode(payload as Uint8Array);
+  const resolvedSeed = typeof payload === "string" ? payload : new TextDecoder().decode(payload);
   cachedWallet = Wallet.fromSeed(resolvedSeed.trim(), { algorithm: ECDSA.ed25519 });
   return cachedWallet;
 }
